backend/ecommerce_app/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
python-jose
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
email-validator
openpyxl
